+++ ndk-build/CHANGELOG.md
@@ -1,3 +1,7 @@
+# Unreleased
+
+- New `ApkConfig` field `apk_name` is now used for APK file naming, instead of the application label.
+
 # 0.2.0 (2021-04-20)
 
 - **Breaking:** refactored `Manifest` into a proper (de)serialization struct. `Manifest` now closely matches [`an android manifest file`](https://developer.android.com/guide/topics/manifest/manifest-element).